requirements.txt
django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
langchain==1.2.0
langchain-core==1.2.4
langchain-openai==1.1.6
langgraph==1.0.5
mysqlclient==2.2.7
openai==2.14.0
pydantic==2.11.9
python-dotenv==1.1.1